Make update-manifest properly update an existing version

+++ src/bin/update-manifest.rs
@@ -5,7 +5,7 @@
 
 use std::io::{stdout, stderr};
 use std::path::PathBuf;
-use std::fs::File;
+use std::fs::{File, OpenOptions};
 use std::error::Error;
 use std::io::prelude::*;
 use std::env;
@@ -37,7 +37,7 @@ fn main () {
     // set the chatpack path variable to the current working directory
     let mut cp_path: PathBuf = env::current_dir().unwrap();
     // then add `target_dir` to that, making `cp_path` the full path to the chatpack directory
-    cp_path.push(TARGET_DIR);
+    cp_path.push(&TARGET_DIR);
     // check if the chatpack directory exists, and return if it doesn't (as there's no work for us to do)
     if !cp_path.exists() {
         println!("The '{}' directory doesn't exist; unable to create / update manifest.", TARGET_DIR);
@@ -47,9 +47,9 @@ fn main () {
     cp_manifest_path.push(&MANIFEST_FILENAME);
     // check if there's an existing hash manifest
     if cp_manifest_path.exists() == false {
-        println!("Creating initial {} manifest...", cp_path.display());
+        println!("Creating initial {} manifest...", TARGET_DIR);
     } else {
-        println!("Rebuilding {}'s manifest...", cp_path.display());
+        println!("Rebuilding {}'s manifest...", TARGET_DIR);
     }
     // clone cp_path (so it's a new object, rather than taking ownership) so cp_version_path can add to it
     let mut cp_version_path = cp_path.clone();
@@ -89,11 +89,43 @@ fn main () {
         Err(why) => panic!("couldn't create or open {}: {}", cp_manifest_path.display(), why.description()),
     } // end of the manifest file open error checking block
     //now update the version        
-    let mut version_file = match File::create(&cp_version_path) {
-        Err(why) => panic!("couldn't create or open {}: {}", cp_version_path.display(), why.description()),
-        Ok(file) => file,
-    };
-    match version_file.write_all(Version::new().to_string().as_bytes()) {
+    // since the lines below are in different scopes (and thus variables defined won't be valid outside of them), I declare the ones I want to keep here
+    let mut version_file;
+    let mut version;
+    // if a version file doesn't exist, create one and populate it with the current version (based on the date)
+    // Otherwise, read what does exist, convert it to a `Version`, and .update() it
+    if cp_version_path.exists() == false {
+        version_file = match File::create(&cp_version_path) {
+            Err(why) => panic!("couldn't create {}: {}", cp_version_path.display(), why.description()),
+            Ok(file) => file,
+        };
+        // get current date-based version, for use further down
+        version = Version::new();
+    } else {
+        // the version file exists, so we need to read it into a string, convert that to a `Version`, run .update(), and then seek to the start of the file so it can be written out
+        version_file = match OpenOptions::new().read(true).write(true).open(&cp_version_path) {
+            Ok(mut file) => {
+                let mut str_ver: String = String::new();
+                // now read all the bytes in the opened file into a string
+                match file.read_to_string(&mut str_ver) {
+                    Ok(_) => (),
+                    Err(why) => panic!("Can't read version file {}: {}", cp_version_path.display(), why.description()),
+                }
+                version = Version::from_string(&str_ver);
+                version.update();
+                // delete the current contents of the file so it can be written to later with just the newly-created version as a string
+                // put the following in a match statement,
+                match file.set_len(0){
+                    Ok(_) => (),
+                    Err(why) => panic!("Can't overwrite {}: {}", cp_version_path.display(), why.description()),
+                }
+                // return version_file
+                file
+            },
+            Err(why) => panic!("Can't write to {}: {}", cp_version_path.display(), why.description()),
+        };
+    } // end the exists if block
+    match version_file.write_all(version.to_string().as_bytes()) {
         Ok(_) => (),
         Err(why) => println!("Can't write new version to {}: {}", cp_manifest_path.display(), why.description()),
     }
